# Don't zero-after-free `DataStream`: ~25% faster IBD