--- a/src/bench/CMakeLists.txt
+++ b/src/bench/CMakeLists.txt
@@ -38,6 +38,7 @@ add_executable(bench_bitcoin
   mempool_eviction.cpp
   mempool_stress.cpp
   merkle_root.cpp
+  net_processing.cpp
   parse_hex.cpp
   peer_eviction.cpp
   poly1305.cpp