--- a/src/bench/rpc_blockchain.cpp
+++ b/src/bench/rpc_blockchain.cpp
@@ -32,8 +32,6 @@ struct TestBlockAndIndex {
     TestBlockAndIndex()
     {
         DataStream stream{benchmark::data::block413567};
-        std::byte a{0};
-        stream.write({&a, 1}); // Prevent compaction
 
         stream >> TX_WITH_WITNESS(block);