--- a/src/support/allocators/zeroafterfree.h
+++ b/src/support/allocators/zeroafterfree.h
@@ -45,7 +45,4 @@ struct zero_after_free_allocator {
     }
 };
 
-/** Byte-vector that clears its contents before deletion. */
-using SerializeData = std::vector<std::byte, zero_after_free_allocator<std::byte>>;
-
 #endif // BITCOIN_SUPPORT_ALLOCATORS_ZEROAFTERFREE_H