--- a/src/test/util/coins.cpp
+++ b/src/test/util/coins.cpp
@@ -8,6 +8,7 @@
 #include <primitives/transaction.h>
 #include <script/script.h>
 #include <test/util/random.h>
+#include <test/util/script.h>
 #include <uint256.h>
 
 #include <stdint.h>
@@ -17,9 +18,9 @@ COutPoint AddTestCoin(FastRandomContext& rng, CCoinsViewCache& coins_view)
 {
     Coin new_coin;
     COutPoint outpoint{Txid::FromUint256(rng.rand256()), /*nIn=*/0};
-    new_coin.nHeight = 1;
+    new_coin.nHeight = rng.rand32();
     new_coin.out.nValue = RandMoney(rng);
-    new_coin.out.scriptPubKey.assign(uint32_t{56}, 1);
+    new_coin.out.scriptPubKey = RandScript(rng, 56);
     coins_view.AddCoin(outpoint, std::move(new_coin), /*possible_overwrite=*/false);
 
     return outpoint;