--- a/src/wallet/wallet.cpp
+++ b/src/wallet/wallet.cpp
@@ -45,7 +45,6 @@
 #include <span.h>
 #include <streams.h>
 #include <support/allocators/secure.h>
-#include <support/allocators/zeroafterfree.h>
 #include <support/cleanse.h>
 #include <sync.h>
 #include <tinyformat.h>