This PR modifies `DataStream`'s vector ` to use the default allocator `std::allocator` rather than `zero_after_free_allocator` which causes a major degradation to performance. The `zero_after_free_allocator` is identical to the default `std::allocator` except that it zeroes memory using `memory_cleanse()` before deallocating.

This PR also drops the `zero_after_free_allocator`, since this was only used by `DataStream` and `SerializeData`, 

In my testing (n=2) on a Raspberry Pi 5 with 4GB of memory, syncing from a fast connection to a stable dedicated node, my branch takes **~74%** of the time taken by master[^1] to sync to height 815,000; average wall clock time was 35h 58m 40s on this branch and 48h 17m 15s on master, see the benchmarking appendix for more detail.

I believe the speedup mainly comes from the use of `DataStream` for all `CDBWrapper` keys and values, and for all of the data we receive from peers. I suspect there are other use cases where performance is improved, but I have not done much testing outside of IBD.

Any objects that contains secrets should *not* be allocated using `zero_after_free_allocator` since they are liable to get mapped to swap space and written to disk if the user is running low on memory, and I intuit this is a likelier path than scanning unzero'd memory for an attacker to find cryptographic secrets. Secrets should be allocated using `secure_allocator` which cleanses on deallocation and `mlock()`s the memory reserved for secrets to prevent it from being mapped to swap space.

## Are any secrets stored in `DataStream` that will lose security?

I have reviewed every appearance of `DataStream` and `SerializeData` as of [`39219fe`](https://github.com/bitcoin/bitcoin/commit/39219fe145e5e6e6f079b591e3f4b5fea8e718040) and have made notes in the appendix below with notes that provide context for each instance where either is used.

The only use case that I wasn't certain of is PSBT's, I believe these are never secrets, but I don't know enough about to know if there are use cases where PSBT's are worthy of being treated as secrets, and being vigilant about not writing them to disk is wise.

As I understand, most of the use of `DataStream` in the wallet code is for the reading and writing of "crypted" key and value data, and they get decrypted somewhere else in a `ScriptPubKeyMan` far away from any `DataStream` container, but I could also be wrong about this, or have misunderstood it's use elsewhere in the wallet.

## Zero-after-free as a buffer overflow mitigation

The `zero_after_free` allocator was added as a buffer overflow mitigation, the idea being that `DataStream`'s store a lot of unsecured data that we don't control like the UTXO set and all P2P messages, and an attacker could fill memory in a predictable way to escalate a buffer overflow into an RCE. (See Historical Background in the Appendix).

I agree completely with practicing security in depth, but I don't think this mitigation is worth the performance hit because: 

1. Aren't there still an abundance of other opportunities for an attacker to fill memory that never gets deallocated?
2. Doesn't ASLR mostly mitigate this issue and don't most devices have some form of ASLR?

I'm not a security expert and I had a hard time finding any writing anywhere that discusses this particular mitigation strategy of zeroing memory, so I hope someone with more knowledge of memory vulnerabilities can assist.

----------

#### Other notes

- I opted to leave `SerializeData` as `std::vector<std::byte>` instead of deleting it and refactoring in the spots where it's used in the wallet to keep the PR small, if others think it would be better to delete it I would be happy to do it.
- I have a feeling that it's not just that we're memsetting everything to 0 in `memory_cleanse` that is causing the performance issue, but the trick we do to prevent compilers from optimizing out the `memset` call is also preventing other optimizations on the `DataStream`'s, but I have yet to test this.
-  I also make a small change to a unit test where boost mysteriously fails to find a left shift-operator for `SerializeData` once it loses its custom allocator.

----------

## Appendices

<details>

<summary>

### Benchmarks

</summary>

Command being timed:
```bash
./src/bitcoind -daemon=0 -connect=amd-ryzen-7900x-node:8333 -stopatheight=815000 -port=8444 -rpcport=8445 -dbcache=2048 -prune=550 -debug=bench -debug=blockstorage -debug=coindb -debug=mempool -debug=prune"
```

I applied my branch on
[6d546336e800](https://github.com/bitcoin/bitcoin/commit/6d546336e800), which is
"master" in the data below.

Average master time (hh:mm:ss): 48:17:15 (173835s)
Average branch time (hh:mm:ss): 35:58:40 (129520s)

~25% reduction in IBD time on a raspberry Pi 5 with a DB cache of 2GB.

#### Master run 1
Wall clock time (hh:mm:ss): 49:38:31 (178711s)

```console
Bitcoin Core version v27.99.0-6d546336e800 (release build)
- Connect block: 158290.53s (620.94ms/blk)
    - Sanity checks: 10.89s (0.01ms/blk)
    - Fork checks: 151.82s (0.02ms/blk)
    - Verify 7077 txins: 135057.68s (165.71ms/blk)
      - Connect 1760 transactions: 134786.36s (165.38ms/blk)
    - Write undo data: 2681.34s (7.38ms/blk)
    - Index writing: 52.76s (0.03ms/blk)
  - Connect total: 138100.75s (611.27ms/blk)
  - Flush: 3933.29s (8.97ms/blk)
  - Writing chainstate: 15814.36s (0.14ms/blk)
  - Connect postprocess: 273.39s (0.52ms/blk)
```

#### Master run 2
Wall clock time (hh:mm:ss): 46:55:58 (168958s)

```
Bitcoin Core version v27.99.0-6d546336e800 (release build)
- Connect block: 145449.95s (940.78ms/blk)
    - Sanity checks: 10.69s (0.01ms/blk)
    - Fork checks: 155.81s (0.02ms/blk)
    - Verify 7077 txins: 115935.55s (142.25ms/blk)
      - Connect 1760 transactions: 115481.15s (141.69ms/blk)
    - Write undo data: 2561.36s (9.05ms/blk)
    - Index writing: 73.63s (0.04ms/blk)
  - Connect total: 118877.56s (929.93ms/blk)
  - Flush: 3864.34s (10.11ms/blk)
  - Writing chainstate: 22294.82s (0.14ms/blk)
  - Connect postprocess: 267.68s (0.56ms/blk)
```

#### Branch run 1
Wall clock time (hh:mm:ss): 34:28:56 (124136s)

```
Bitcoin Core version v27.99.0-a0dddf8b4092 (release build)
- Connect block: 107134.59s (1017.01ms/blk)
    - Sanity checks: 11.01s (0.01ms/blk)
    - Fork checks: 150.93s (0.03ms/blk)
    - Verify 7077 txins: 87446.53s (107.30ms/blk)
      - Connect 1760 transactions: 87329.99s (107.15ms/blk)
    - Write undo data: 2495.47s (7.36ms/blk)
    - Index writing: 37.95s (0.04ms/blk)
  - Connect total: 90318.60s (1006.42ms/blk)
  - Flush: 3917.28s (9.92ms/blk)
  - Writing chainstate: 12560.43s (0.15ms/blk)
  - Connect postprocess: 259.89s (0.47ms/blk)
```

#### Branch run 2
Wall clock time (hh:mm:ss): 37:28:24 (134904s)

```
Bitcoin Core version v27.99.0-a0dddf8b4092 (release build)
- Connect block: 117991.55s (144.77ms/blk)
  - Connect total: 101298.20s (124.29ms/blk)
    - Sanity checks: 11.17s (0.01ms/blk)
    - Fork checks: 151.24s (0.19ms/blk)
    - Verify 7077 txins: 98446.38s (120.79ms/blk)
      - Connect 1760 transactions: 98339.79s (120.66ms/blk)
    - Write undo data: 2484.75s (3.05ms/blk)
    - Index writing: 36.62s (0.04ms/blk)
  - Flush: 3892.28s (4.78ms/blk)
  - Writing chainstate: 12446.33s (15.27ms/blk)
  - Connect postprocess: 259.11s (0.32ms/blk)
```
</details>

<details>

<summary>

### Historical background

</summary>

At some point prior to the oldest git commit for the repo, an allocator `secure_allocator` was [added](https://github.com/bitcoin/bitcoin/blob/0a61b0df1224a5470bcddab302bc199ca5a9e356/serialize.h#L675-L702) that zeroes out memory on deallocation with `memset()`, and was [used](https://github.com/bitcoin/bitcoin/blob/0a61b0df1224a5470bcddab302bc199ca5a9e356/serialize.h#L711-L714) as the allocator for the vector `vch` in `CDataStream` (now `DataStream`).

In July 2011, PR [#352](https://github.com/bitcoin/bitcoin/pull/352) adding support for encrypted wallets `secure_allocator` was [modified](https://github.com/bitcoin/bitcoin/pull/352/commits/c1aacf0be347b10a6ab9bbce841e8127412bce41) to also `mlock()` data on allocation to prevent the wallet passphrase or other secrets from being paged to swap space (written to disk).

In January 2012, findings were shared (https://bitcointalk.org/index.php?topic=56491.0) that [#352](https://github.com/bitcoin/bitcoin/pull/352) modifying `CDataStream`'s allocator slowed down IBD substantially[^2], since `CDataStream` was used in many places that did not need the guarantees of `mlock()`, and since every call to `mlock()` results in a flush of the TLB (a cache that maps virtual memory to physical memory).

PR [#740](https://github.com/bitcoin/bitcoin/pull/740) was opened to fix this, initially[^3] by removing the custom allocator `secure_allocator` from `CDataStream`'s `vector_type`:

```diff
 class CDataStream
 {
 protected:
-    typedef std::vector<char, secure_allocator<char> > vector_type;
+    typedef std::vector<char> vector_type;
     vector_type vch;
```

A reviewer of [#740](https://github.com/bitcoin/bitcoin/pull/740) [suggested](https://github.com/bitcoin/bitcoin/pull/740#issuecomment-3356239) that dropping `mlock()` was a good idea, but that the original behavior of zeroing-after-freeing (should it be zeroing-*before*-freeing?) `CDataStream` should be restored as a mitigation for buffer overflows:

> I love the performance improvement, but I still don't like the elimination of zero-after-free. Security in depth is important.
>
> Here's the danger:
>
> Attacker finds a remotely-exploitable buffer overrun somewhere in the networking code that crashes the process.
> They turn the crash into a full remote exploit by sending carefully constructed packets before the crash packet, to initialize used-but-then-freed memory to a known state.
>
> Unlikely? Sure.
>
> Is it ugly to define a zero_after_free_allocator for CDataStream? Sure. (simplest implementation: copy secure_allocator, remove the mlock/munlock calls).
>
> But given that CDataStream is the primary interface between bitcoin and the network, I think being extra paranoid here is a very good idea.

Another reviewer benchmarked `CDataStream` with an allocator that zeroed memory using `memset` without `mlock`ing it and found that performance was almost identical to the default allocator, while both were substantially faster than the `mlock`ing variant of `CDataStream`. (https://web.archive.org/web/20130622160044/https://people.xiph.org/~greg/bitcoin-sync.png).

Based on the benchmark, and the potential security benefit, the `zero_after_free` allocator was created and used as `CDataStream`'s allocator.

In November 2012, PR [#1992](https://github.com/bitcoin/bitcoin/pull/1992) was opened to address the fact that in many cases `memset()` calls are optimized away by compilers as part of a family of compiler optimizations called [dead store elimination](https://www.usenix.org/conference/usenixsecurity17/technical-sessions/presentation/yang) by replacing the `memset` call with openssl's `OPENSSL_cleanse` which is meant to solve this problem. Given that all of the data being zero'ed out in the deallocator is also having it's only pointer destroyed, these memset calls were candidates for being optimized.

I suspect that the reason no performance regression was found in the benchmarking of [#740](https://github.com/bitcoin/bitcoin/pull/740) which introduced the `zero_after_free` allocator is that the `memset` calls were being optimized out.

I am not the first to suggest that this is a performance issue:

https://bitcoin-irc.chaincode.com/bitcoin-core-dev/2015-11-06#1446837840-1446854100;

https://bitcoin-irc.chaincode.com/bitcoin-core-dev/2016-11-23#1479883620-1479882900;

Or to write a patch changing it:

https://github.com/bitcoin/bitcoin/commit/671c724716abdd69b9d253a01f8fec67a37ab7d7

</details>

<details>

<summary>

### All uses of DataStream and SerializeData
( ⚠️ when opening: very long)

</summary>

I performed this review on commit [39219fe145e5e6e6f079b591e3f4b5fea8e71804](https://github.com/bitcoin/bitcoin/commit/39219fe145e5e6e6f079b591e3f4b5fea8e71804)

I look, briefly, at every single use of `DataStream` outside of test code, to see whether or not it contains secret information that should be zeroed out, or should be mlocked to prevent paging to swap.

I've taken liberties to editorialize some of the codeblocks below for legibility, and all comments that have `[]` are my own.

##### `DataStream`

In `src/addrdb.cpp`+`src/addrdb.h`:

```cpp
/** Only used by tests. */
void ReadFromStream(AddrMan& addr, DataStream& ssPeers);
```

Only used by tests.

-----

In `src/addrman.cpp` `Addrman::Serialize(DataStream&)` & `Unserialize(DataStream&)`, are explicitly instantiated, these are used in `SerializeFileDB` and `DeserializeDB` which are used to serialize (`DumpPeerAddresses`) addrman to disk, and to deserialize addrman from disk (`LoadAddrman`).

The most valuable secret seems to be addrman's `nKey` used to determine the address buckets randomly.

-------

In `src/blockencodings.cpp`:

```cpp
void CBlockHeaderAndShortTxIDs::FillShortTxIDSelector() const {
    DataStream stream{};
    stream << header << nonce;
    CSHA256 hasher;
    hasher.Write((unsigned char*)&(*stream.begin()), stream.end() - stream.begin());
    uint256 shorttxidhash;
    hasher.Finalize(shorttxidhash.begin());
    shorttxidk0 = shorttxidhash.GetUint64(0);
    shorttxidk1 = shorttxidhash.GetUint64(1);
}
```

Here we are just using the DataStream to be able to Serialize the block header and nonce into a string of bytes that get hashed to make short id k0 and k1 for [BIP 152](https://github.com/bitcoin/bips/blob/master/bip-0152.mediawiki#short-transaction-ids).

This gets invoked when we construct a `CBlockHeaderandShortTxIDs` for an INV of type `MSG_CMPCT_BLOCK` in `PeerManagerImpl::SendMessage()`.

-------

In `src/common/blooms.cpp`:

DataStream is used to deserialize outpoints into our bloom filter, these are not secrets in any way:

```cpp
void CBloomFilter::insert(const COutPoint& outpoint)
{
    DataStream stream{};
    stream << outpoint;
    insert(MakeUCharSpan(stream));
}
```

-------

In `src/core_read.cpp`:

DataStream is used in `DecodeTx` for serialization/deserialization of the transaction data, used afaict only in RPC's for deserializing user arguments into `CMutableTransaction`'s.

It's used in `DecodeHexBlockHeader()`which deserializes a block header argument into a `CBlockHeader` for the `submitheader` rpc.

Similar for `DecodeHexBlk()` used by the `getblocktemplate` and `submitblock` rpc's.

----

In `src/core_write.cpp`:

```cpp
void CBloomFilter::insert(const COutPoint& outpoint)
{
    DataStream stream{};
    stream << outpoint;
    insert(MakeUCharSpan(stream));
}
```

`EncodeHexTx` is only used in RPC's, and transaction data does not contain secrets.

------

In `dbwrapper.h` and `dbwrapper.cpp` it is used exclusively to serialize and deserialize coinsdb keys and values, none of which is secret.

--------

In `src/external_signer`:

```cpp
bool ExternalSigner::SignTransaction(PartiallySignedTransaction& psbtx, std::string& error)
{
    // Serialize the PSBT
    DataStream ssTx{};
    ssTx << psbtx;
```

I don't think this is a secret, but I don't know enough about PSBT's to be sure.

-------

There is some scaffolding for being able to transmit serializable stuff over the IPC wire in `src/capnp/common-types.h`, I assume this depends on how it's used, nothing essentially secret.

--------

In `src/kernel/coinstats.cpp`:

```cpp
void ApplyCoinHash(MuHash3072& muhash, const COutPoint& outpoint, const Coin& coin)
{
    DataStream ss{};
    TxOutSer(ss, outpoint, coin);
    muhash.Insert(MakeUCharSpan(ss));
}
```

Here it's used for serializing oupoints and coins for creating the AssumeUTXO assumed utxo set hash, nothing secret.

-------

In `src/net.cpp`:

In `ConvertSeeds()` serialized seeds get converted into usable address objects, we initialize a DataStream with the input seeds that we are going to try connecting to during node bootstrapping.

```cpp
//! Convert the serialized seeds into usable address objects.
static std::vector<CAddress> ConvertSeeds(const std::vector<uint8_t> &vSeedsIn)
{
    // It'll only connect to one or two seed nodes because once it connects,
    // it'll get a pile of addresses with newer timestamps.
    // Seed nodes are given a random 'last seen time' of between one and two
    // weeks ago.
    const auto one_week{7 * 24h};
    std::vector<CAddress> vSeedsOut;
    FastRandomContext rng;
    ParamsStream s{DataStream{vSeedsIn}, CAddress::V2_NETWORK};
    while (!s.eof()) {
        CService endpoint;
        s >> endpoint;
        CAddress addr{endpoint, SeedsServiceFlags()};
        addr.nTime = rng.rand_uniform_delay(Now<NodeSeconds>() - one_week, -one_week);
        LogDebug(BCLog::NET, "Added hardcoded seed: %s\n", addr.ToStringAddrPort());
        vSeedsOut.push_back(addr);
    }
    return vSeedsOut;
}
```

It is also used for creating an empty `CNetMessage` which has a `DataStream` member in `CNetMessage V2Transport::GetReceivedMessage()`:

```cpp
//! Convert the serialized seeds into usable address objects.
static std::vector<CAddress> ConvertSeeds(const std::vector<uint8_t> &vSeedsIn)
{
    // It'll only connect to one or two seed nodes because once it connects,
    // it'll get a pile of addresses with newer timestamps.
    // Seed nodes are given a random 'last seen time' of between one and two
    // weeks ago.
    const auto one_week{7 * 24h};
    std::vector<CAddress> vSeedsOut;
    FastRandomContext rng;
    ParamsStream s{DataStream{vSeedsIn}, CAddress::V2_NETWORK};
    while (!s.eof()) {
        CService endpoint;
        s >> endpoint;
        CAddress addr{endpoint, SeedsServiceFlags()};
        addr.nTime = rng.rand_uniform_delay(Now<NodeSeconds>() - one_week, -one_week);
        LogDebug(BCLog::NET, "Added hardcoded seed: %s\n", addr.ToStringAddrPort());
        vSeedsOut.push_back(addr);
    }
    return vSeedsOut;
}
```

--------

In `net.h`

`CNetMessage` the universal p2p message container used a `DataStream` to store received message data.

```cpp
/** Transport protocol agnostic message container.
 * Ideally it should only contain receive time, payload,
 * type and size.
 */
class CNetMessage
{
public:
    DataStream m_recv;                   //!< received message data
    std::chrono::microseconds m_time{0}; //!< time of message receipt
    uint32_t m_message_size{0};          //!< size of the payload
    uint32_t m_raw_message_size{0};      //!< used wire size of the message (including header/checksum)
    std::string m_type;

    explicit CNetMessage(DataStream&& recv_in) : m_recv(std::move(recv_in)) {}
    // Only one CNetMessage object will exist for the same message on either
    // the receive or processing queue. For performance reasons we therefore
    // delete the copy constructor and assignment operator to avoid the
    // possibility of copying CNetMessage objects.
    CNetMessage(CNetMessage&&) = default;
    CNetMessage(const CNetMessage&) = delete;
    CNetMessage& operator=(CNetMessage&&) = default;
    CNetMessage& operator=(const CNetMessage&) = delete;
};
```

It's also used for the lower level handling of messages, including partially received header buffers and received socket data in `V1Transport` as in v2 transport above in `net.cpp`.

```cpp
/** Transport protocol agnostic message container.
 * Ideally it should only contain receive time, payload,
 * type and size.
 */
…